Migrate away from Gio
This replaces all uses of Gio with Jello (for rendering), curve (for
types), and our own wsi (for window management).

Support for pointer events is currently missing.

+++ widget/flex.go
@@ -45,7 +45,7 @@ func (f *Flex) CreateElement() Element {
 }
 
 type Flexible struct {
-	Flex  float32
+	Flex  float64
 	Fit   render.FlexFit
 	Child Widget
 }
